generalize default error
missing dot

missing dot
