+++ packages/next/src/client/components/dev-root-http-access-fallback-boundary.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React from 'react'
-import { HTTPAccessFallbackBoundary } from './http-access-fallback-boundary'
+import { HTTPAccessFallbackBoundary } from './http-access-fallback/error-boundary'
 
 export function bailOnRootNotFound() {
   throw new Error('notFound() is not allowed to use in root layout')
